backend/requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.10
django-cors-headers==3.10.1
django-import-export==2.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
psycopg2
psycopg2-binary
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==4.0.1
xlrd==2.0.1
xlwt==1.3.0
